Promote mounts to spec
 On-behalf-of: @SAP mangirdas.judeikis@sap.com
