+++ test/e2e/mounts/crd_kubecluster.yaml
